refactor(tools): extract notes domain into src/tools/notes.py
