+++ src/tools/__init__.py
@@ -22,3 +22,4 @@ from src.tools.cookbook import (  # noqa: F401
     _MODEL_PROCESS_PATTERNS,
 )
 from src.tools.search import do_search_chats  # noqa: F401
+from src.tools.notes import do_manage_notes  # noqa: F401
